Performace tuning : - Tuner reported AiInterface::findtarget to be the biggest bottleneck. Old version used 0.249% of total process time in a 5 minute test. New version uses 0.040% process time. That means a rough 600% improvement. All gains have a backfire !!. Do not jump and use this ref on big servers. Wait until people can test and confirm it does not have problems. Problems that could ocure : - mobs attacking stuff that should not - mobs have a big delay to start attacking on summoned non player units - small crash probability (lol, this should not happen but somehow it always pops up :D )
!!this is only biggest bottleneck. This does not solve world hunger and the lack of wet tee shirt contests.

+++ src/ascent-world/MiscHandler.cpp
@@ -1559,7 +1559,7 @@ void WorldSession::HandleInspectOpcode( WorldPacket & recv_data )
 
 	m_Packed_GUID.appendPackGUID( player->GetGUID());
 	uint32 guid_size;
-	guid_size = m_Packed_GUID.size();
+	guid_size = (uint32)m_Packed_GUID.size();
 
 	data.append(m_Packed_GUID);
 	data << uint32( talent_points );
